+++ src/main/java/net/silentchaos512/powerscale/setup/PsLoot.java
@@ -7,9 +7,7 @@ import net.neoforged.neoforge.common.loot.IGlobalLootModifier;
 import net.neoforged.neoforge.registries.DeferredRegister;
 import net.neoforged.neoforge.registries.NeoForgeRegistries;
 import net.silentchaos512.powerscale.PowerScale;
-import net.silentchaos512.powerscale.loot.condition.CrystalDropChanceConfig;
-import net.silentchaos512.powerscale.loot.condition.CrystalDropWithCooldown;
-import net.silentchaos512.powerscale.loot.condition.MobProperties;
+import net.silentchaos512.powerscale.loot.condition.*;
 import net.silentchaos512.powerscale.loot.modifier.BonusDropsLootModifier;
 
 import java.util.function.Supplier;
@@ -22,6 +20,14 @@ public class PsLoot {
 
     // Conditions
 
+    public static final Supplier<LootItemConditionType> BLIGHT_CRYSTAL_DROPS_CONFIG = LOOT_CONDITION_TYPES.register(
+            "blight_crystal_drops_config",
+            () -> new LootItemConditionType(BlightCrystalDropConfig.CODEC)
+    );
+    public static final Supplier<LootItemConditionType> BOSS_CRYSTAL_DROPS_CONFIG = LOOT_CONDITION_TYPES.register(
+            "boss_crystal_drops_config",
+            () -> new LootItemConditionType(BossCrystalDropConfig.CODEC)
+    );
     public static final Supplier<LootItemConditionType> CRYSTAL_DROP_CHANCE_CONFIG = LOOT_CONDITION_TYPES.register(
             "crystal_drop_chance_config",
             () -> new LootItemConditionType(CrystalDropChanceConfig.CODEC)